+++ packages/module-source/src/module-source.js
@@ -55,21 +55,21 @@ const analyzeModule = makeModuleAnalyzer();
  * @property {SourceMapHook} [sourceMapHook]
  */
 
-// XXX implements import('ses').PrecompiledStaticModuleInterface but adding
+// XXX implements import('ses').PrecompiledModuleSource but adding
 // `@implements` errors that this isn't a class and `@returns` errors that
 // there's no value returned.
 /**
- * StaticModuleRecord captures the effort of parsing and analyzing module text
- * so a cache of StaticModuleRecords may be shared by multiple Compartments.
+ * ModuleSource captures the effort of parsing and analyzing module text
+ * so a cache of ModuleSources may be shared by multiple Compartments.
  *
  * @class
  * @param {string} source
  * @param {string | Options} [opts]
  */
-export function StaticModuleRecord(source, opts = {}) {
+export function ModuleSource(source, opts = {}) {
   if (new.target === undefined) {
     throw TypeError(
-      "Class constructor StaticModuleRecord cannot be invoked without 'new'",
+      "Class constructor ModuleSource cannot be invoked without 'new'",
     );
   }
   if (typeof opts === 'string') {