+++ packages/ses/src/error/assert.js
@@ -566,3 +566,7 @@ export { makeAssert };
 /** @type {Assert} */
 const assert = makeAssert();
 export { assert };
+
+// Internal, to obviate polymorphic dispatch, but may become rigorously
+// consistent with @endo/error.
+export { makeError, note as annotateError, redactedDetails as X, quote as q };